requirements/requirements.txt
django==4.2.7
python-dotenv==1.0.0
djangorestframework==3.14.0
pytest-django==4.5.2
requests-mock==1.11.0
pillow==10.1.0
django-model-utils==4.3.1
